# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: tanmaytambat01@gmail.com
    name: API Support
  description: A cloud infrastructure autoscaling service that monitors cluster metrics
    and automatically scales resources based on demand.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cloud Autoscaler API
  version: "1.0"
paths:
...
  /clusters/{id}/events:
    get:
      description: Get scaling events for a cluster
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: from
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: to
        type: string
      - description: Relative time range (e.g., 1h, 24h, 7d)
        in: query
        name: range
        type: string
      - default: 50
        description: Maximum number of results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Scaling events
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cluster not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get scaling events
      tags:
      - Events
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"